# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for live stream configuration, ingest/publish lifecycle, DVR recordings, and event hooks.",
        "title": "Open Streamer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/streams": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streams"
                ],
                "summary": "List streams",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.StreamList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/streams/{code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streams"
                ],
                "summary": "Get stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stream code (a-zA-Z0-9_)",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.StreamData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streams"
                ],
                "summary": "Create or partial-update stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stream code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial or full stream document",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apidocs.StreamPutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.StreamData"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apidocs.StreamData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "streams"
                ],
                "summary": "Delete stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stream code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/streams/{code}/inputs/switch": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streams"
                ],
                "summary": "Manual input switch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stream code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target input priority",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apidocs.InputSwitchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.StreamActionData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/streams/{code}/restart": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streams"
                ],
                "summary": "Restart stream pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stream code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.StreamActionData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apidocs.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "apidocs.InputSwitchRequest": {
            "type": "object",
            "properties": {
                "priority": {
                    "type": "integer"
                }
            }
        },
...
        "apidocs.StreamActionData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/apidocs.StreamActionInner"
                }
            }
        },
...
        "apidocs.StreamData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/apidocs.StreamResponse"
                }
            }
        },
        "apidocs.StreamList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apidocs.StreamResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "apidocs.StreamPutRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "dvr": {
                    "$ref": "#/definitions/domain.StreamDVRConfig"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Input"
                    }
                },
                "name": {
                    "type": "string"
                },
                "protocols": {
                    "$ref": "#/definitions/domain.OutputProtocols"
                },
                "push": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PushDestination"
                    }
                },
                "stream_key": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "$ref": "#/definitions/domain.ThumbnailConfig"
                },
                "transcoder": {
                    "$ref": "#/definitions/domain.TranscoderConfig"
                },
                "watermark": {
                    "$ref": "#/definitions/domain.WatermarkConfig"
                }
            }
        },
...
        "domain.StreamDVRConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "max_size_gb": {
                    "description": "MaxSizeGB caps total disk usage. Oldest segments pruned when exceeded.\n0 = no limit.",
                    "type": "number"
                },
                "retention_sec": {
                    "description": "RetentionSec is the retention window in seconds.\n0 = keep forever.",
                    "type": "integer"
                },
                "segment_duration": {
                    "description": "SegmentDuration overrides the global segment length in seconds.\n0 = use default (4s).",
                    "type": "integer"
                },
                "storage_path": {
                    "description": "StoragePath overrides the default DVR root directory for this stream.\n\"\" = use \"./dvr/{streamCode}\".",
                    "type": "string"
                }
            }
        },
...
        "domain.ThumbnailConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "height": {
                    "type": "integer"
                },
                "interval_sec": {
                    "description": "IntervalSec generates one thumbnail every N seconds.",
                    "type": "integer"
                },
                "output_dir": {
                    "description": "OutputDir is relative to the publisher HLS directory.\nE.g. \"thumbnails\" → written to {hls_dir}/{stream_code}/thumbnails/thumb.jpg",
                    "type": "string"
                },
                "quality": {
                    "description": "Quality is the JPEG quality (1–31, lower = better). Default: 5.",
                    "type": "integer"
                },
                "width": {
                    "description": "Width and Height of the output thumbnail in pixels.\n0 = match source resolution.",
                    "type": "integer"
                }
            }
        },
...
    }
}